+++ GrappleEditor/src/GrappleEditor/EditorLayer.cpp
@@ -214,7 +214,7 @@ namespace Grapple
             {
                 if (resolutions[i] == settings.Resolution)
                 {
-                    resolutionIndex = i;
+                    resolutionIndex = (uint32_t)i;
                     break;
                 }
             }
